requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
dataclasses-json==0.6.1
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.310
langchain-experimental==0.0.27
langsmith==0.0.43
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
numpy==1.24.4
openai==0.28.1
packaging==23.2
pandas==2.0.3
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
joblib
fastapi==0.103.2
yarl==1.9.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
uvicorn==0.22.0
scipy
Flask
Flask-Cors
gdown
python-dotenv
retry
-e .
